+++ src/Distance_sensor/GP2D12.cpp
@@ -31,7 +31,7 @@ void GP2D12_Innit(int pinNumber)
  * @return float
  * distance to object (in centimeters)
  */
-unsigned short GP2D12_Read(int pinNumber)
+float GP2D12_Read(int pinNumber)
 {
 
     // tension lu par l'arduino 10cm = 2.45V environ,40cm = 0.75V, 50cm = 0.62V environ et 80cm 0.4V
